\name{noNA}
\alias{noNA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Gastric cancer case-control study
}
\description{Example dataset for the SvyNom package.
}
\usage{data(noNA)
}
\author{
Mithat Gonen, Marinela Capanu
}
\examples{data(noNA)
}
\keyword{survival}
\keyword{survey}
\keyword{nomogram}
