\name{SvyNom-package}
\alias{SvyNom-package}
\alias{SvyNom}
\docType{package}
\title{Nomograms for right-censored outcomes with complex survey data}
\description{
Builds, evaluates and validates a nomogram with survey data and right-censored outcomes.
}
\details{
\tabular{ll}{
Package: \tab SvyNom\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-01-06\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}
There are three functions for the user
svycox.nomogram, svycox.validate, svycox.calibrate
}
\author{
Marinela Capanu and Mithat Gonen \cr
Maintainer: gonenm@mskcc.org
}
\references{
Marinela Capanu, Mithat Gonen (2015). Building a Nomogram for Survey-Weighted Cox Models Using R. 
Journal of Statistical Software, Code Snippets, 64(1), 1-17. URL https://www.jstatsoft.org/v64/c01/.
}
\keyword{ package }
