\name{SvyNom-package}
\alias{SvyNom-package}
\alias{SvyNom}
\docType{package}
\title{Nomograms with Righ-Censored Survey Data}
\description{
Builds, evaluates and validates a nomogram with survey data and right-censored outcomes
}
\details{
\tabular{ll}{
Package: \tab SvyNom\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-02\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
There are three functions for the user
svycox.nomogram, svycox.validate, svycok.calibrate
}
\author{
Marinela Capanu and Mithat Gonen
Maintainer: gonenm@mskcc.org
}
\references{
Capanu M and Gonen M (2012). J Stat Soft
}
\keyword{ package }
